We add a filter by order id and now we can update a order.
The class OrderDetailPage have a final parameter of OrderPageBloc type to update the list of orders that we showed on the first search. When we create a new order, we add it to the list, and when we updated it we replace in the list of the stream called "orders" in the OrderPageBloc.
We add a new attribute to the Order model called dateApproved. That will update when we are approving a order.
We create the service to update a order.
Changes of look and feel in the orders.
We can't change the creation date of the order.
We change the label of the employed who making the order from "Aplicante" to "Solicitante".