We validate the form to create and update an order.
We call to the changed service (by the back-end) to get all the branches by user, when the user doesn't have a role with code '01'. Now we don't wait for a initial search, the query can be empty.
We fixed some issues with the name of the labels in the order form.