Fixing issues with the login page and order's detail. And we add a name to the app.
We have to control better the error messages in the login stage.
Now we fetch the access by role from the login or validating the user logged.
We have to change the type of the detail's quantity from 'int' to 'double'.
Because we are not using an int in the order's detail, we have to change the way to update this field. We can't use the icons buttons 'add' and 'remove' anymore. Instead, now we use a TextField to fill the data.
We add a listener to show a snackbar with the quantity of orders fetched.
We add the name to the app, its name is FiBo Pedidos.