Navigation changes.
We implemented a router to control the navigation by the data base and the user's access.
We add 2 new attributes to the program model, 'path' for the router and 'icon' for the drawer.
Now when we call a new page we use the function to replace all the pages stacked in the navigation.
We control the navigation  to disabled the page if we are currently in it.
A class ConversionDataType to control a specifics functions of conversion between data types. In this case we control when change the string aspect of a double type.