Fixing keyboard issue in some device.
When we want to write in a TextFormField, the keyboard is come up, and it hidden the persistentFooterButtons from the Scaffold, so we add an action button in the AppBar to hide the keyboard. It's added to other pages where we have to use the keyboard.
We change the colors of the orders approved, from blueAccent to green.