Order state definition was bad.
We had to change the behavior of the pages by the order's state. The user pass us the right definition; (A) Pending, (P) Processed and (X) Canceled.
Only if the user create the order he could edit itself. We had to make a change in the StreamBuilder of the OrderPageDetail to control that.
When the user "process" an order now it have to add a "commentary". We add the new attribute in the class.
We have to create a new page to let the user add the commentary to process the order.