Icon have been incorporated.
Version 0.1.5+1
In the services of the models where we search a list of them, we change the called, now we send the parameter "name" in the GET request, and not in the URL.
Order detail page, we have add the "toUpperCase()" method to the validation user created and user logged, because we have issues with this validation when the user in the back end is in minuscules.
In the data searches, we call the function to search even if the query is empty.